Move more modules to services for organization
