+++ api/v1/controllers/class/end.js
@@ -2,7 +2,7 @@ const { hasClassPermission } = require("@middleware/permission-check");
 const { isAuthenticated } = require("@middleware/authentication");
 const { endClass } = require("@services/class-service");
 const { CLASS_PERMISSIONS } = require("@modules/permissions");
-const { classStateStore } = require("@modules/classroom");
+const { classStateStore } = require("@services/classroom-service");
 
 module.exports = (router) => {
     /**
