+++ api/v1/controllers/room/student.js
@@ -1,5 +1,5 @@
 const { isAuthenticated } = require("@middleware/authentication");
-const { classStateStore } = require("@modules/classroom");
+const { classStateStore } = require("@services/classroom-service");
 const AuthError = require("@errors/auth-error");
 
 module.exports = {
