+++ app.js
@@ -26,6 +26,7 @@ const NotFoundError = require("@errors/not-found-error");
 const { logout } = require("@services/user-service");
 const { passport } = require("@modules/google-oauth.js");
 const { rateLimiter } = require("@middleware/rate-limiter");
+const { ensureFormbarDeveloperPool } = require("@services/bootstrap-service");
 
 // Create session for user information to be transferred from page to page
 const sessionMiddleware = session({
@@ -199,6 +200,12 @@ app.use(errorHandlerMiddleware);
 // Start the server
 
 http.listen(settings.port, async () => {
+    try {
+        await ensureFormbarDeveloperPool();
+    } catch (err) {
+        console.error("Failed to ensure Formbar Developer Pool exists:", err);
+    }
+
     // Object.assign(authentication.whitelistedIps, await getIpAccess("whitelist"));
     // Object.assign(authentication.blacklistedIps, await getIpAccess("blacklist"));
     console.log(`Running on port: ${settings.port}`);
