+++ sockets/middleware/rate-limiter.js
@@ -1,4 +1,4 @@
-const { PASSIVE_SOCKETS } = require("@modules/socket-updates");
+const { PASSIVE_SOCKETS } = require("@services/socket-updates-service");
 const { TEACHER_PERMISSIONS } = require("@modules/permissions");
 const { handleSocketError } = require("@modules/socket-error-handler");
 const { socketStateStore } = require("@stores/socket-state-store");
